Updated LDoA bot
updated LDoA
